Refactor security functions into SecurityHelper class
Introduced a new `SecurityHelper` class to centralize password hashing and link sterilization functionalities. Updated various command handlers and controllers to replace calls to the old `Helper` methods with the new `SecurityHelper` methods, improving code organization and security practices. Adjusted using directives to include necessary namespaces for the new class.
