+++ src/Moonglade.Auth/ValidateLoginCommand.cs
@@ -19,7 +19,7 @@ public class ValidateLoginCommandHandler(
         if (account is null) return Task.FromResult(false);
         if (account.Username != request.Username) return Task.FromResult(false);
 
-        var valid = account.PasswordHash == Helper.HashPassword(request.InputPassword.Trim(), account.PasswordSalt);
+        var valid = account.PasswordHash == SecurityHelper.HashPassword(request.InputPassword.Trim(), account.PasswordSalt);
 
         if (!valid)
         {
