+++ src/Moonglade.Web/Controllers/SettingsController.cs
@@ -213,7 +213,7 @@ public class SettingsController(
             }
 
             // Sterilize
-            link.Url = Helper.SterilizeLink(link.Url);
+            link.Url = SecurityHelper.SterilizeLink(link.Url);
         }
 
         blogConfig.SocialLinkSettings = new()
@@ -285,14 +285,14 @@ public class SettingsController(
     [ProducesResponseType(StatusCodes.Status409Conflict)]
     public async Task<IActionResult> UpdateLocalAccountPassword(UpdateLocalAccountPasswordRequest request)
     {
-        var oldPasswordValid = blogConfig.LocalAccountSettings.PasswordHash == Helper.HashPassword(request.OldPassword.Trim(), blogConfig.LocalAccountSettings.PasswordSalt);
+        var oldPasswordValid = blogConfig.LocalAccountSettings.PasswordHash == SecurityHelper.HashPassword(request.OldPassword.Trim(), blogConfig.LocalAccountSettings.PasswordSalt);
 
         if (!oldPasswordValid) return Conflict("Old password is incorrect.");
 
-        var newSalt = Helper.GenerateSalt();
+        var newSalt = SecurityHelper.GenerateSalt();
         blogConfig.LocalAccountSettings.Username = request.NewUsername.Trim();
         blogConfig.LocalAccountSettings.PasswordSalt = newSalt;
-        blogConfig.LocalAccountSettings.PasswordHash = Helper.HashPassword(request.NewPassword, newSalt);
+        blogConfig.LocalAccountSettings.PasswordHash = SecurityHelper.HashPassword(request.NewPassword, newSalt);
 
         await SaveConfigAsync(blogConfig.LocalAccountSettings);
         return NoContent();
