+++ src/Moonglade.Web/Pages/Components/SocialLink/Default.cshtml
@@ -9,7 +9,7 @@
         <div role="list">
             @foreach (var item in Model.OrderBy(c => c.Name))
             {
-                <a href="@Helper.SterilizeLink(item.Url)" target="_blank" class="d-block mb-3 mt-2" role="listitem">
+                <a href="@SecurityHelper.SterilizeLink(item.Url)" target="_blank" class="d-block mb-3 mt-2" role="listitem">
                     <i class="@item.Icon me-1"></i>
                     @item.Name
                 </a>
