+++ src/Moonglade.Web/Pages/Post.cshtml
@@ -135,7 +135,7 @@
     @if (!string.IsNullOrWhiteSpace(Model.Post.HeroImageUrl))
     {
         <div class="post-hero-image-container text-center mb-3">
-            <img src="@Helper.SterilizeLink(Model.Post.HeroImageUrl)" class="post-hero-image img-fluid rounded-3 shadow-sm" />
+            <img src="@SecurityHelper.SterilizeLink(Model.Post.HeroImageUrl)" class="post-hero-image img-fluid rounded-3 shadow-sm" />
         </div>
     }
